alterado tipo campo valor

+++ database/migrations/2022_11_30_155534_create_servicos_table.php
@@ -17,8 +17,10 @@ return new class extends Migration
             $table->id();
             $table->string('nome');
             $table->string('descricao');
-            $table->integer('valor_cliente')->default(0);
-            $table->integer('valor_profissional')->default(0);
+            $table->decimal('valor_cliente', 5, 2)->default(0);
+            $table->decimal('valor_profissional', 5, 2)->default(0);
+            //$table->integer('valor_cliente')->default(0);
+            //$table->integer('valor_profissional')->default(0);
             $table->string('unidade');
             $table->string('horario');
             $table->string('periodo_minimo');
